lb_token
Module execute

Functions

exec_change_balance 🔒

change token balance of an existing token_id.

exec_curate_token_id

checks if token_id is available (ie: not yet created), then creates new token_id and initial balances

impl_send 🔒

Implements a single Send function. Transfers Uint256 amount of a single token_id, saves transfer history, may register-receive, and creates callback message.

impl_transfer 🔒

Implements a single Transfer function. Transfers a Uint256 amount of a single token_id and saves the transfer history. Used by Transfer and Send (via impl_send) messages

is_valid_name 🔒

is_valid_symbol 🔒

pad_response

try_add_receiver_api_callback 🔒

try_batch_send

try_batch_transfer

try_burn_tokens

try_change_admin

try_change_metadata

try_create_viewing_key

try_give_permission

does not check if token_id exists so attacker cannot easily figure out if a token_id has been created

try_mint_tokens

try_register_receive

try_remove_admin

try_revoke_permission

changes an existing permission entry to default (ie: revoke all permissions granted). Does not remove entry in storage, because it is unecessarily in most use cases, but will require also removing owner-specific PermissionKeys, which introduces complexity and increases gas cost. If permission does not exist, message will return an error.

try_revoke_permit

try_send

try_set_viewing_key

try_transfer

verify_admin 🔒

verify_curator 🔒

verifies if sender is a curator

verify_minter 🔒

verifies if sender is a minter of the specific token_id