lb_token::execute
Function exec_change_balance

fn exec_change_balance(
    storage: &mut dyn Storage,
    token_id: &str,
    remove_from: Option<&Addr>,
    add_to: Option<&Addr>,
    amount: &Uint256,
    token_info: &StoredTokenInfo,
) -> StdResult<()>

change token balance of an existing token_id.

Should check that token_id already exists before calling this function, which is not done explicitly in this function. Although token_info is an argument in this function, so it is likely that the calling function would have had to check.

If remove_from == None: minted new tokens.

If add_to == None: burn tokens.

If is_nft == true, then remove_from MUST be Some(_).

If is_nft == true, stores new owner of NFT