lb_token::state::permissions
Function append_permission_for_addr

fn append_permission_for_addr(
    storage: &mut dyn Storage,
    owner: &Addr,
    token_id: &str,
    allowed_addr: &Addr,
) -> StdResult<()>

stores a PermissionKey {token_id: String, allowed_addr: String] for a given owner. Note that permission key is [owner, token_id, allowed_addr]. This function does not enforce that the list of PermissionKey stored is unique; while this doesn’t really matter, the ref implementation’s functions aim to ensure each entry is unique, for storage efficiency.