lb_token::state
Module save_load_functions

Functions

json_load

Returns StdResult from retrieving the item with the specified key using Json (de)serialization because bincode2 annoyingly uses a float op when deserializing an enum.
Returns a StdError::NotFound if there is no item with that key

json_save

Returns StdResult<()> resulting from saving an item to storage using Json (de)serialization because bincode2 annoyingly uses a float op when deserializing an enum