lb_token::state::save_load_functions
Function json_save

pub fn json_save<T: Serialize>(
    storage: &mut dyn Storage,
    key: &[u8],
    value: &T,
) -> StdResult<()>

Returns StdResult<()> resulting from saving an item to storage using Json (de)serialization because bincode2 annoyingly uses a float op when deserializing an enum

Arguments

storage - a mutable reference to the storage this item should go to

key - a byte slice representing the key to access the stored item

value - a reference to the item to store